fixed HTML errors and warnings
The HTML validator threw an error about image tags not having "alt" so I have added the description for each image. Also, it warned that under the section tag, it needs to have either headings or div so I have encapsulated  all the content under section with "div" tag. It still warns but we can ignore the warnings now.